+++ src/models/world/query.rs
@@ -5,7 +5,7 @@ use diesel::{QueryDsl, QueryResult, RunQueryDsl};
 
 pub async fn get_character_max_by_account_and_world_id(
     state: SharedState,
-    acc_id: i64,
+    acc_id: i32,
     world_id: i16,
 ) -> QueryResult<i32> {
     let db = {
